refactor(permissions): move GLOBAL/PermissionScope to scopes submodule
Re-export the GLOBAL scope and the PermissionScope class from a new
app.lib.permissions.scopes submodule instead of the app.lib.permissions package facade,
mirroring how app.lib.permissions.registry is imported directly. The facade no longer
surfaces them; callers and tests now import them from app.lib.permissions.scopes.

Update CLAUDE.md, README, and the plugin guide so the documented import surface matches.
