+++ app/lib/permissions/scopes.py
@@ -1,0 +1,6 @@
+from app.core.permissions.scopes import GLOBAL, PermissionScope
+
+__all__ = [
+    "PermissionScope",
+    "GLOBAL",
+]
